Fix skin::Submesh::indexStart truncation, dedupe alternate-texture embeds, close out BLENDER_EXPORT_TODO.md
The real bug behind "missing body geometry" on large character models:
Submesh::Level is the high 16 bits of indexStart, not LOD metadata as
assumed and left unread across three earlier investigation passes --
confirmed against reference/wow.export's own Skin.js. Any .skin whose
resolved triangle-index buffer exceeds 65,535 entries (77/114 submeshes
on a real bloodelffemale_hd00.skin) was silently aliased into the wrong
region of the buffer. indexStart widened to uint32_t, computed as
(level << 16) | rawIndexStart.

Also: --collision is now opt-in (was opt-out via --collision none) after
the collision hull was found occluding a real character render by
default. Ambiguous hardcoded-texture-slot candidates (2+ same-basename
files) are now all embedded as alternate_textures extras instead of
silently dropped -- and deduplicated by filename across materials after
that feature caused a real 1.9GB/5m39s export (19 materials sharing one
94-file pool, each re-embedding it independently); fixed to 104MB/5.4s.

EYES_ON_FINDINGS.md and BLENDER_EXPORT_TODO.md updated with the full
investigation trails, including wrong turns that got corrected along the
way. BLENDER_EXPORT_TODO.md deleted -- every item it tracked is now
resolved. FILE_SPLIT_TODO.md also deleted (its own splits were already
fully implemented, just undocumented as such).

+++ src/skin.hpp
@@ -34,8 +34,8 @@ struct Header {
 // M2SkinSection, per wowdev.wiki M2/.skin#Submeshes -- 48 bytes on disk.
 // `skinSectionId` (the "Mesh part ID"/"geoset ID") plus the fields needed
 // to slice this submesh's triangles out of the skin's flat index buffer are
-// surfaced; `Level`/centerPosition/sortCenterPosition/sortRadius/boneCount/
-// boneComboIndex/boneInfluences/centerBoneIndex stay unread (LOD/culling/
+// surfaced; centerPosition/sortCenterPosition/sortRadius/boneCount/
+// boneComboIndex/boneInfluences/centerBoneIndex stay unread (culling/
 // skinning-optimization concerns, not materials). `skinSectionId` is what a
 // real client uses to decide *which* submeshes to actually draw for a given
 // character/creature configuration -- a `.skin` file routinely bundles every
@@ -46,11 +46,32 @@ struct Header {
 // `cmd_export.cpp`'s `buildMaterialsAndPrimitives`, which surfaces every
 // distinct value actually used as a loud note rather than silently merging
 // them with no indication multiple geosets got exported unfiltered.
+//
+// `Level` (offset 0x02, between skinSectionId and vertexStart) is NOT
+// LOD/culling metadata despite the field name and wowdev.wiki's own
+// placement of it near those concerns -- confirmed against a real, working
+// independent implementation (wow.export's Skin.js): it's the *high 16
+// bits* of `indexStart` (real client code: `triangleStart += level << 16`).
+// A .skin's on-disk `indexStart`/`triangleStart` field is only 16 bits, so
+// any model whose resolved triangle-index buffer exceeds 65,535 entries
+// (real, not rare -- confirmed on `bloodelffemale_hd.m2`'s own 136,254-
+// entry buffer, where 77 of 114 submeshes need this) needs `Level` to
+// address into the back half of that buffer at all. `indexStart` here is
+// already the corrected 32-bit value (`(level << 16) | rawIndexStart`),
+// computed once in `parseSubmeshes` -- callers never need to know `Level`
+// exists. Found the hard way: an earlier session read the field name,
+// assumed "per-LOD selection marker" without checking what a real client
+// does with it, and left it unread -- for any submesh needing correction,
+// husk silently sliced the *wrong* region of the triangle-index buffer
+// (aliased into the first 65,536 entries) instead of throwing or visibly
+// failing, so this went undetected until a real interactive Blender
+// render showed missing body geometry and the investigation traced all
+// the way back here.
 struct Submesh {
     uint16_t skinSectionId = 0;
     uint16_t vertexStart = 0;
     uint16_t vertexCount = 0;
-    uint16_t indexStart = 0;  // into the skin's resolved triangle-index buffer
+    uint32_t indexStart = 0;  // into the skin's resolved triangle-index buffer; see Level note above
     uint16_t indexCount = 0;
 };
